refactor(orchestration): name inject's rejection message for what it is
Rename the module/function to orchestration-inject-rejection-message /
buildInjectRejectionMessage, cut the six-line block comment to two WHY lines,
drop the unreachable trim/filter over a static config, and fold the duplicated
dispatch assertion into the existing rejection test (now pinned to the builder
output). Trims the recognition test to the real command lines it locks. No
behavior change.
